refactor(mentor):  items 페이지에 멘토링시간에 배운 분리 적용 (presentational & container pattern)
